# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "一个简单的机器管理系统",
        "title": "rabbit",
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/api/v1/tag/all": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "全部tag数据",
                "parameters": [
                    {
                        "type": "string",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "orderBy",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.APIGetTagListOutputs"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/tag/create": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "创建新tag",
                "parameters": [
                    {
                        "description": "创建新tag",
                        "name": "APIPostTagCreateInputs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.APIPostTagCreateInputs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.APIPostTagCreateInputs"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/tag/info": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "tag 详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tag id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.G6Edge"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    }
                }
            }
        },
        "/api/v1/tag/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "tag列表",
                "parameters": [
                    {
                        "type": "string",
                        "name": "categoryID",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "orderBy",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "remark",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.APIGetTagListOutputs"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/tag/update": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "summary": "更新tag",
                "parameters": [
                    {
                        "description": "更新tag",
                        "name": "APIPostTagCreateInputs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.APIPostTagCreateInputs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.APIPostTagCreateInputs"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/tag_category/all": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "tag category列表",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "orderBy",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.APIGetTagCategoryListOutputs"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/tag_category/create": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "创建tag category",
                "parameters": [
                    {
                        "description": "创建tag category",
                        "name": "APIPostTagCategoryCreateInputs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.APIPostTagCategoryCreateInputs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.TagCategory"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/tag_category/info": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "tag category 详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tag category id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.TagCategory"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/tag_category/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "tag category列表",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "orderBy",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.APIGetTagCategoryListOutputs"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/tag_category/tags": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "查看某一tag category下的tag",
                "parameters": [
                    {
                        "type": "string",
                        "name": "categoryID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "categoryName",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.APIGetTagCategoryTagsOutputs"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/tag_category/update": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "summary": "更新tag category",
                "parameters": [
                    {
                        "description": "更新tag category",
                        "name": "APIPostTagCategoryCreateInputs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.APIPostTagCategoryCreateInputs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.TagCategory"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app.APIGetTagCategoryListOutputs": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.TagCategory"
                    }
                },
                "totalCount": {
                    "type": "integer"
                }
            }
        },
...
        "app.APIGetTagCategoryTagsOutputs": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.Tag"
                    }
                },
                "totalCount": {
                    "type": "integer"
                }
            }
        },
...
        "app.APIGetTagListOutputs": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.Tag"
                    }
                },
                "totalCount": {
                    "type": "integer"
                }
            }
        },
...
        "app.APIPostTagCategoryCreateInputs": {
            "type": "object",
            "properties": {
                "categoryID": {
                    "type": "integer"
                },
                "cnName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                }
            }
        },
        "app.APIPostTagCreateInputs": {
            "type": "object",
            "properties": {
                "categoryID": {
                    "type": "integer"
                },
                "cnName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                }
            }
        },
...
        "app.G6Edge": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "object",
                    "$ref": "#/definitions/app.G6Point"
                },
                "endPoint": {
                    "type": "object",
                    "$ref": "#/definitions/app.G6Point"
                },
                "id": {
                    "type": "string"
                },
                "shape": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "sourceId": {
                    "description": "edge.sourceID 指向 node.id",
                    "type": "integer"
                },
                "start": {
                    "type": "object",
                    "$ref": "#/definitions/app.G6Point"
                },
                "startPoint": {
                    "type": "object",
                    "$ref": "#/definitions/app.G6Point"
                },
                "target": {
                    "type": "string"
                },
                "targetId": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "app.G6Point": {
            "type": "object",
            "properties": {
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                }
            }
        },
        "app.Tag": {
            "type": "object",
            "properties": {
                "categoryID": {
                    "type": "integer"
                },
                "categoryName": {
                    "type": "string"
                },
                "cnName": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "createAt": {
                    "type": "object",
                    "$ref": "#/definitions/gtime.GTime"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "inPoints": {
                    "type": "string"
                },
                "isDoingEnd": {
                    "type": "boolean"
                },
                "isDoingStart": {
                    "type": "boolean"
                },
                "label": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "outPoints": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "shape": {
                    "type": "string"
                },
                "size": {
                    "type": "string"
                },
                "stateImage": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "x": {
                    "type": "integer"
                },
                "y": {
                    "type": "integer"
                }
            }
        },
        "app.TagCategory": {
            "type": "object",
            "properties": {
                "cnName": {
                    "type": "string"
                },
                "createAt": {
                    "type": "object",
                    "$ref": "#/definitions/gtime.GTime"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                }
            }
        },
...
        "gtime.GTime": {
            "type": "object"
        },
...
    }
}